Add service startname to output
The user context which the service is running under was added to the
output of Get-ServiceUnquoted, Get-ServiceFilePermission, and
Get-ServicePermission. This is useful in situations where many services
were found to be vulnerable on the local machine however only a few are
running under the context of a different user (preferably NT
AUTHORITY\SYSTEM or a member of Domain Admins group).