Add service DACL functionality
Added 2 helper functions: Get-ServiceDaclPermission and
Test-ServiceDaclPermission. For this pull request, these functions are
used in Invoke-ServiceStart, Invoke-ServiceStop, Invoke-ServiceEnable,
and Invoke-ServiceDisable in an attempt to ensure that service
permissions are checked using the "correct" method. However, the plan is
to use these functions for further code cleanup, refactoring, and
additional features in the near future.